refactor(queue): 대기열 도메인 패키지를 core-domain으로 이동
