+++ core/core-api/src/test/java/com/ticket/core/api/controller/QueueControllerSecurityTest.java
@@ -1,11 +1,11 @@
 package com.ticket.core.api.controller;
 
 import com.ticket.core.config.LoginMemberArgumentResolver;
-import com.ticket.core.domain.member.MemberPrincipal;
-import com.ticket.core.domain.queue.usecase.GetQueueStatusUseCase;
-import com.ticket.core.domain.queue.usecase.ExitQueueUseCase;
-import com.ticket.core.domain.queue.usecase.JoinQueueUseCase;
-import com.ticket.core.enums.Role;
+import com.ticket.core.config.security.MemberPrincipal;
+import com.ticket.core.domain.queue.command.exit.ExitQueueUseCase;
+import com.ticket.core.domain.queue.command.join.JoinQueueUseCase;
+import com.ticket.core.domain.queue.query.status.GetQueueStatusUseCase;
+import com.ticket.core.domain.member.model.Role;
 import com.ticket.core.support.ApiControllerAdvice;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
