--- core/core-api/src/test/java/com/ticket/core/domain/queue/usecase/QueueEntryIdTest.java
@@ -1,26 +1,0 @@
-package com.ticket.core.domain.queue.usecase;
-
-import com.ticket.core.support.exception.CoreException;
-import com.ticket.core.support.exception.ErrorType;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-@SuppressWarnings("NonAsciiCharacters")
-class QueueEntryIdTest {
-
-    @Test
-    void 앞뒤_공백을_제거한다() {
-        QueueEntryId queueEntryId = QueueEntryId.from("  qe-1  ");
-
-        assertThat(queueEntryId.value()).isEqualTo("qe-1");
-    }
-
-    @Test
-    void 빈값이면_예외를_던진다() {
-        assertThatThrownBy(() -> QueueEntryId.from("   "))
-                .isInstanceOf(CoreException.class)
-                .satisfies(error -> assertThat(((CoreException) error).getErrorType()).isEqualTo(ErrorType.INVALID_REQUEST));
-    }
-}
