+++ core/core-domain/src/main/java/com/ticket/core/domain/queue/model/QueueLevel.java
