+++ core/core-domain/src/main/java/com/ticket/core/domain/queue/runtime/QueueTicket.java
