+++ core/core-domain/src/test/java/com/ticket/core/domain/queue/runtime/QueueTicketTest.java
